Moved files around, added annotated route with notes
