+++ routes.js
@@ -1,0 +1,7 @@
+module.exports = function(app) {
+     
+    app.get('/', function(req,res) {
+         res.send("Hello Boring World!");
+    });     
+    
+};
