+++ server.js
@@ -2,13 +2,13 @@
 var express    =    require('express');
 var app        =    express();
 
-require('./router/main')(app);
+require('./routes')(app);
 app.set('views',__dirname + '/views');
 app.set('view engine', 'ejs');
 app.engine('html', require('ejs').renderFile);
 
-var server     =    app.listen(3000,function(){
-console.log("Express is running on port 3000");
+var server     =    app.listen(8080,function(){
+console.log("Express is running on port 8080");
 });
