+++ views/index.html
@@ -1,6 +1,6 @@
 
 <body>
 
-<div> HI, <%= phrase %> </div>
+<div> <%= phrase %> </div>
 
 <body>
